app/requirements.txt
pandas==2.2.3
psycopg2_binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.2
PyYAML==6.0.2
Requests==2.32.3
